@@ -3,16 +3,18 @@ package com.goesplayer;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.design.widget.FloatingActionButton;
-import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.AdapterView;
 import android.widget.ListView;
 import android.widget.TextView;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+
+import com.google.android.material.floatingactionbutton.FloatingActionButton;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -70,9 +72,11 @@ public class FolderFragment extends Fragment {
 
     private void criarLista() {
         ArrayList<String> listAux = new ArrayList<>();
-        for (int i = 0; i < MainActivity.todasMusicas.size(); i++) listAux.add(MainActivity.todasMusicas.get(i).getFolder());
+        for (int i = 0; i < MainActivity.todasMusicas.size(); i++)
+            listAux.add(MainActivity.todasMusicas.get(i).getFolder());
         folderList = new ArrayList<>(new HashSet<>(listAux));
-        for (int i = 0; i < folderList.size(); i++) if (folderList.get(i).equals("<unknown>")) folderList.remove(i);
+        for (int i = 0; i < folderList.size(); i++)
+            if (folderList.get(i).equals("<unknown>")) folderList.remove(i);
         Collections.sort(folderList, new Comparator<String>() {
             @Override
             public int compare(String o1, String o2) {
